Allow dynamic ports to be assigned with loopback address
When using the -p 0 option, ListenLocalHost cannot be used. Unfortunately specifying `-o 0 -a 127.0.0.1` also does not work
since ShouldUseLocalhost returns true for anything that resembles a local loopback address.

This change uses the equivalent API that allows port 0 to be specified. In addition, it consistently normalizes to the loopback
address both when printing the "listening on" message and when launching a browser